Move OSCWalkingGains to another header file
