+++ examples/Cassie/osc/BUILD.bazel
@@ -7,6 +7,7 @@ cc_library(
     deps = [
         "//examples/Cassie/osc:heading_traj_generator",
         "//examples/Cassie/osc:high_level_command",
+        "//examples/Cassie/osc:osc_walking_gains",
         "//examples/Cassie/osc:standing_com_traj",
         "//examples/Cassie/osc:standing_pelvis_traj",
         "//examples/Cassie/osc:walking_speed_control",
@@ -74,6 +75,14 @@ cc_library(
         "//lcmtypes:lcmt_robot",
         "//systems/controllers:control_utils",
         "//systems/framework:vector",
+        "@drake//:drake_shared_library",
+    ],
+)
+
+cc_library(
+    name = "osc_walking_gains",
+    hdrs = ["osc_walking_gains.h"],
+    deps = [
         "@drake//:drake_shared_library",
     ],
 )
